Shallow git repo: optimize update to avoid fetching every time
`git fetch` with shallow remotes isn't very smart, and re-fetches stuff
that we already have; re-implement the functionality we need in a
smarter way. Get the updated ref (if necessary) through `git ls-remote`,
and try to checkout the relevant commit; only if this fails, fall back
to `git fetch`ing it.