Allow the root not to have a configured remote
the root wasn't ever required to be under a VCS, but still has a
"project" object in the dependency tree for internal reasons; if it's
actually under git, but doesn't have an origin remote (yet?), currently
quark blows up, because the root repo is recognized as a git repo, but
the `git remote get-url origin` call that is automatically done fails.

Notice that this kind of failure was already expected and ignored in
`create_dependency_tree`, but it expected a `QuarkError`, while here we
get a "raw" CalledProcessError.

To fix this, in case of failure of the command above we translate the
error to a `QuarkError` instead, so that the error is not swallowed, but
in this context is correctly ignored.