Fix foreach.t cram test to cope with newer Git versions
Newer Git versions produce additional warnings when init.defaultBranch
is unset.

Since it's generally unset, we set to some value for the test's
execution, so that we don't get unexpected output and we don't force
developers to fiddle with yet another Git configuration knob when they
want to run tests locally.