Add "quark run"
This adds each submodule's "bin" directory to the path and will make it
easier to consume development tools that are downloaded by Quark (e.g.
by using the "devdepends" table in `subprojects.quark` to e.g. download
and use a project-specific version of CMake).

The directory itself is configurable via the `quark_run_paths` option
that accepts a list of module-relative paths that should be considered.
Its default value is `["bin"]`, this way most packages will work out of
the box by following the existing convention of putting executables in a
`bin` directory.

The current implementation currently considers all modules in the tree,
in order to keep things simple (otherwise we'd have to add provenance
information to all dependencies which would make this patch a bit more
invasive).

In the future we may want to consider only "devdepends" dependencies but
we would like to play with this feature in real world scenarios first.